findSVI: Calculate Social Vulnerability Index for Communities

Developed by CDC/ATSDR (Centers for Disease Control and Prevention/ Agency for Toxic Substances and Disease Registry), Social Vulnerability Index (SVI) serves as a tool to assess the resilience of communities by taking into account socioeconomic and demographic factors. Provided with year(s), region(s) and a geographic level of interest, 'findSVI' retrieves required variables from US census data and calculates SVI for communities in the specified area based on CDC/ATSDR SVI documentation. Reference for the calculation methods: Flanagan BE, Gregory EW, Hallisey EJ, Heitgerd JL, Lewis B (2011) doi:10.2202/1547-7355.1792 .

See also

Author

Maintainer: Heli Xu xuheli91@gmail.com (ORCID) [copyright holder]

Authors:

Ran Li rl627@drexel.edu (ORCID)

Usama Bilal ub45@drexel.edu (ORCID)