Retrieve Census Data and Calculate Social Vulnerability Index (SVI) for One or Multiple Year(s)/State(s)

This is a wrapper for get_census_data() and get_svi() that retrieves census data and produces SVI for one or multiple years(s) and state(s). For multiple year-state entries, SVI is obtained from entry-specific percentile ranking and summarised into one table.

Usage

find_svi(year, state = NULL, geography, key = NULL, full.table = FALSE)

Arguments

year

A vector containing years of interest (available 2012-2021). Length >=1. Acting as pairs with state, year should be of the same length as state. The exception is when it's a single year entry (length 1), other than providing one state of interest, supply state = NULL as default or state = 'US' retrieves and processes nation level data to obtain SVI.

state

A vector containing states of interest. Length >=0. Length 0 (state = NULL), or state = 'US' must be used with single year argument, when SVI is calculated from nation-level census data. In other cases, state must have the same elements as year (same length).

geography

One geography level of interest for all year-state combination (e.g."county", "zcta", "tract").

key

Your Census API key. Obtain one at https://api.census.gov/data/key_signup.html. To set up, use census_api_key("YOUR KEY GOES HERE"), or include it as an argument.

full.table

Default as FALSE, returning SVI table with only "GEOID", and SVI for each theme and all themes. If set as TRUE, a full SVI table with individual SVI variables and intermediate ranking calculations are also included in addition to the theme-related SVIs (similar style to tables from CDC/ATSDR database).

Value

A tibble of summarised SVI for one or multiple year-state combination of interest. Rows represent the geographic units, and columns represent its SVI for each theme and all themes. Additional two columns at the end indicate the corresponding state and year information. For full.table = TRUE, estimated count and percentage values for individual SVI variables are also included. For description of variable names (column names), please refer to CDC/ATSDR documentation.

Examples

if (FALSE) {
census_api_key("YOUR KEY GOES HERE")

# Use with vectors for year and state
## for one year-state information (all ZCTAs of PA in 2019)
find_svi(
      year = 2019,
      state = "PA",
      geography = "tract"
   )

## for multiple year-state combination (all ZCTAs of RI in 2017 and PA in 2018)
find_svi(
      year = c(2017, 2018),
      state = c("RI", "PA"),
      geography = "zcta"
   )

# Use with a table of year-state information
info <- tribble(
    ~state, ~year,
    "AZ", 2015,
    "AZ", 2016,
    "PA", 2020,
    "RI", 2018)

find_svi(
    year = info$year,
    state = info$state,
    geography = "county"
   )
}